+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrMetroFactory.kt
@@ -8,7 +8,6 @@ import dev.zacsweers.metro.compiler.ir.parameters.Parameters
 import dev.zacsweers.metro.compiler.ir.parameters.parameters
 import dev.zacsweers.metro.compiler.memoize
 import org.jetbrains.kotlin.ir.builders.IrBuilderWithScope
-import org.jetbrains.kotlin.ir.builders.irGetObject
 import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrConstructor
 import org.jetbrains.kotlin.ir.declarations.IrFunction
@@ -17,12 +16,9 @@ import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.symbols.IrFunctionSymbol
 import org.jetbrains.kotlin.ir.types.typeWith
 import org.jetbrains.kotlin.ir.util.TypeRemapper
-import org.jetbrains.kotlin.ir.util.companionObject
 import org.jetbrains.kotlin.ir.util.deepCopyWithSymbols
 import org.jetbrains.kotlin.ir.util.defaultType
 import org.jetbrains.kotlin.ir.util.hasAnnotation
-import org.jetbrains.kotlin.ir.util.isFromJava
-import org.jetbrains.kotlin.ir.util.isObject
 import org.jetbrains.kotlin.ir.util.remapTypes
 import org.jetbrains.kotlin.ir.util.simpleFunctions
 import org.jetbrains.kotlin.name.Name
@@ -46,13 +42,7 @@ internal sealed interface IrMetroFactory {
   ): IrExpression =
     with(scope) {
       // Anvil may generate the factory
-      val isJava = factoryClass.isFromJava()
-      val creatorClass =
-        if (isJava || factoryClass.isObject) {
-          factoryClass
-        } else {
-          factoryClass.companionObject()!!
-        }
+      val creatorClass = factoryClass.requireStaticIshDeclarationContainer()
       val createFunction = creatorClass.simpleFunctions().first { it.name in createFunctionNames }
 
       val remapper = createFunction.typeRemapperFor(typeKey.type)
@@ -73,12 +63,7 @@ internal sealed interface IrMetroFactory {
 
       val args = computeArgs(finalFunction, parameters)
       val createExpression =
-        irInvoke(
-          dispatchReceiver = if (isJava) null else irGetObject(creatorClass.symbol),
-          callee = createFunction.symbol,
-          args = args,
-          typeHint = factoryClass.typeWith(),
-        )
+        irInvoke(callee = createFunction.symbol, args = args, typeHint = factoryClass.typeWith())
 
       // Wrap in a metro provider if this is a provider
       return if (isDaggerFactory && factoryClass.defaultType.implementsProviderType()) {